+++ coreClasses/gameMenu/labels/ArrowLabel.h
@@ -15,7 +15,7 @@ using ArrowLabel_sptr = std::shared_ptr<ArrowLabel>;
 
 class ArrowLabel : public Label {
 public:
-    ArrowLabel(Node_sptr node, const JBTypes::Color &color, float id);
+    ArrowLabel(Node_sptr node, const JBTypes::Color &color, int id);
 
     ~ArrowLabel() override = default;
